+++ src/common/route-criteria.js
@@ -1,10 +1,24 @@
-// Route criteria object.
+// Route criteria objects.
 
-
-// Constructor.
-function defineRouteCriteria(crNumber, crSign)
+const criteriaTypesEnum =
 {
-	var defineRes = {"number": crNumber, "sign": crSign};
+	"STOP_COUNT": 1,
+	"TOTAL_DISTANCE": 2
+};
+
+
+// 'Stop Count' constructor.
+function defineStopCountCriteria(inpCount, inpSign)
+{
+	var defineRes = setNumberSign(criteriaTypesEnum.STOP_COUNT, inpCount, inpSign);
+	return defineRes;
+}
+
+
+// 'Total Distance' constructor.
+function defineTotalDistanceCriteria(inpDist, inpSign)
+{
+	var defineRes = setNumberSign(criteriaTypesEnum.TOTAL_DISTANCE, inpDist, inpSign);
 	return defineRes;
 }
 
@@ -19,7 +33,7 @@ function validateRouteCriteria(givenObject)
 	if (givenObject !== null && valueType === "object")
 	{
 		// Check properties.
-		validationResult = handleCriteriaObject(givenObject);
+		validationResult = handleNumberSign(givenObject);
 	}
 	else if (givenObject !== null)
 	{
@@ -37,7 +51,7 @@ function validateRouteCriteria(givenObject)
 
 
 // Check object properties.
-function handleCriteriaObject(criteriaObj)
+function handleNumberSign(criteriaObj)
 {
 	var numberVal = criteriaObj.number;
 	var correctNumType = Number.isInteger(numberVal);
@@ -55,8 +69,22 @@ function handleCriteriaObject(criteriaObj)
 }
 
 
+function setNumberSign(typeVal, numVal, signVal)
+{
+	var setRes = {};
+	
+	setRes["type"] = typeVal;
+	setRes["number"] = numVal;
+	setRes["sign"] = signVal;
+	
+	return setRes;
+}
+
+
 module.exports =
 {
-	defineCriteria: defineRouteCriteria,
-	validateCriteria: validateRouteCriteria,
+	criteriaTypes: criteriaTypesEnum,
+	defineStopCount: defineStopCountCriteria,
+	defineTotalDistance: defineTotalDistanceCriteria,
+	validateCriteria: validateRouteCriteria
 };